Added formatting for all numeric outputs
Added formatting for all numeric outputs in the
ProductionReportController.  If the output is currency it is formatted
with a $ sign and to 2 decimal places.  If the output is not currency it
is formatted to 2 decimal places.  Material variance is formatted as a
percentage.  

Changed the way that materialVariance is calculated in the
ProductionEntry class.